Removed some more compiler warnings in retroshare-gui.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@4566 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/RemoteDirModel.cpp
@@ -761,7 +761,7 @@ QModelIndex FlatStyle_RDM::index(int row, int column, const QModelIndex & parent
 	if(row < 0)
 		return QModelIndex() ;
 
-	if(row < _ref_entries.size())
+	if(row < (int) _ref_entries.size())
 	{
 		void *ref = _ref_entries[row].first ;
